+++ templates/workspaces/base/terraform/azure-monitor/azure-monitor.tf
@@ -180,20 +180,71 @@ resource "azurerm_private_endpoint" "azure_monitor_private_endpoint" {
     is_manual_connection           = false
   }
 
-  private_dns_zone_group {
-    name = "azure-monitor-private-dns-zone-group"
+  depends_on = [
+    azurerm_monitor_private_link_scoped_service.ampls_app_insights,
+    azurerm_monitor_private_link_scoped_service.ampls_log_anaytics,
+  ]
+}
 
-    private_dns_zone_ids = [
-      var.azure_monitor_dns_zone_id,
-      var.azure_monitor_oms_opinsights_dns_zone_id,
-      var.azure_monitor_ods_opinsights_dns_zone_id,
-      var.azure_monitor_agentsvc_dns_zone_id,
-      var.blob_core_dns_zone_id,
-    ]
+# Separate DNS zone group using azapi to avoid AnotherOperationInProgress errors
+# See: https://github.com/hashicorp/terraform-provider-azurerm/issues/28715
+resource "azapi_resource" "azure_monitor_dns_zone_group" {
+  type      = "Microsoft.Network/privateEndpoints/privateDnsZoneGroups@2023-11-01"
+  name      = "azure-monitor-private-dns-zone-group"
+  parent_id = azurerm_private_endpoint.azure_monitor_private_endpoint.id
+
+  body = {
+    properties = {
+      privateDnsZoneConfigs = [
+        {
+          name = "privatelink-monitor-azure-com"
+          properties = {
+            privateDnsZoneId = var.azure_monitor_dns_zone_id
+          }
+        },
+        {
+          name = "privatelink-oms-opinsights-azure-com"
+          properties = {
+            privateDnsZoneId = var.azure_monitor_oms_opinsights_dns_zone_id
+          }
+        },
+        {
+          name = "privatelink-ods-opinsights-azure-com"
+          properties = {
+            privateDnsZoneId = var.azure_monitor_ods_opinsights_dns_zone_id
+          }
+        },
+        {
+          name = "privatelink-agentsvc-azure-automation-net"
+          properties = {
+            privateDnsZoneId = var.azure_monitor_agentsvc_dns_zone_id
+          }
+        },
+        {
+          name = "privatelink-blob-core-windows-net"
+          properties = {
+            privateDnsZoneId = var.blob_core_dns_zone_id
+          }
+        }
+      ]
+    }
+  }
+
+  response_export_values = ["id"]
+
+  retry = {
+    error_message_regex  = ["AnotherOperationInProgress", "Another operation on this or dependent resource is in progress"]
+    interval_seconds     = 15
+    max_interval_seconds = 90
+  }
+
+  timeouts {
+    create = "30m"
+    delete = "30m"
   }
 
   depends_on = [
-    azurerm_monitor_private_link_scoped_service.ampls_app_insights,
+    azurerm_private_endpoint.azure_monitor_private_endpoint,
   ]
 }