+++ CHANGELOG.md
@@ -10,6 +10,9 @@ ENHANCEMENTS:
 * Add support for setting resource processor VMSS SKU via environment variables ([#4936](https://github.com/microsoft/AzureTRE/issues/4936))
 * Exclude recovery service vaults from e2e tests ([#4920](https://github.com/microsoft/AzureTRE/issues/4920))
 
+BUG FIXES:
+* Fix intermittent workspace deletion failure caused by `AnotherOperationInProgress` 409 errors on the Azure Monitor (AMPLS) private DNS zone group, by adding an azapi delete retry. ([#3194](https://github.com/microsoft/AzureTRE/issues/3194))
+
 ## (0.28.0) (March 2, 2026)
 **BREAKING CHANGES**
 * Sonatype Nexus shared service now requires explicit EULA acceptance (`accept_nexus_eula: true`) when deploying. This ensures compliance with Sonatype Nexus Community Edition licensing. ([#4842](https://github.com/microsoft/AzureTRE/issues/4842))